+++ CHANGES.rst
@@ -11,6 +11,8 @@ Unreleased
   comments or submissions
 * :class:`.RedditorStream` has been added to facilitate
   :attr:`.Redditor.stream`
+* :meth:`.Inbox.collapse` to mark messages as collapsed.
+* :meth:`.Inbox.uncollapse` to mark messages as uncollapsed.
 
 **Fixed**